+++ src/pages/RollingPaperList/RollingPaperList.page.jsx
@@ -1,23 +1,9 @@
-import React, { useEffect, useState } from 'react';
-import recipientService from '../../api/services/recipients.services';
+import React from 'react';
+import useRecipients from '../../components/common/hooks/useRecipients';
 
 function RollingPaperList() {
-  const [rollingPapers, setRollingPapers] = useState();
-
-  async function getRecipients() {
-    const res = await recipientService.getRecipients();
-    const { results, count } = res.data;
-    console.log(results);
-    return results;
-  }
-
-  useEffect(() => {
-    const results = getRecipients();
-    setRollingPapers(results);
-  }, []);
-
+  const { rollingPapers } = useRecipients();
   console.log(rollingPapers);
-
   return <div>RollingPaperList</div>;
 }
